Kenux大更新

+++ kernel/arch/x86_64/include/arch/slab.h
@@ -52,8 +52,8 @@ void kmem_cache_free(kmem_cache_t* cache, void* obj);
 void* slab_alloc(uint64_t size);
 void slab_free(void* ptr);
 
-void* kmalloc(uint64_t size);
-void* kzalloc(uint64_t size);
+void* kmalloc(size_t size);
+void* kzalloc(size_t size);
 void  kfree(void* ptr);
 uint64_t ksize(void* ptr);
